requirements_amm.txt
web3>=6.9.0
python-dotenv>=1.0.0
